requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anyio @ file:///C:/ci_311/anyio_1676425491996/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci_311/argon2-cffi-bindings_1676424443321/work
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
attrs @ file:///C:/b/abs_35n0jusce8/croot/attrs_1695717880170/work        
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.7.0
Brotli==1.1.0
build==1.1.1
cachetools==5.3.2
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dash.ly==0.17.3
debugpy==1.8.0
decorator==5.1.1
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work     
Deprecated==1.2.14
distro==1.9.0
entrypoints @ file:///C:/ci_311/entrypoints_1676423328987/work
et-xmlfile==1.1.0
executing==2.0.0
fastapi==0.110.0
fastjsonschema @ file:///C:/ci_311/python-fastjsonschema_1679500568724/work
filelock==3.13.1
Flask==2.1.3
Flask-Compress==1.14
Flask-SeaSurf==1.1.1
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
halo==0.0.31
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna @ file:///C:/ci_311/idna_1676424932545/work
importlib-metadata==6.11.0
importlib_resources==6.1.2
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
joblib==1.2.0
jsonschema @ file:///C:/b/abs_d40z05b6r1/croot/jsonschema_1678983446576/work
jupyter-server @ file:///C:/ci_311/jupyter_server_1678228762759/work      
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server_proxy==4.1.0
jupyterlab-dash==0.1.0a3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keras==2.15.0
kiwisolver==1.4.4
kubernetes==29.0.0
libclang==16.0.6
log-symbols==0.0.14
lxml @ file:///C:/b/abs_9e7tpg2vv9/croot/lxml_1695058219431/work
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune @ file:///C:/ci_311/mistune_1676425149302/work
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work      
mrjob==0.7.4
multidict==6.0.4
nbclassic @ file:///C:/b/abs_c8_rs7b3zw/croot/nbclassic_1681756186106/work
nbclient @ file:///C:/ci_311/nbclient_1676425195918/work
nbconvert @ file:///C:/ci_311/nbconvert_1676425836196/work
nbformat @ file:///C:/b/abs_5a2nea1iu2/croot/nbformat_1694616866197/work  
nest-asyncio==1.5.8
notebook @ file:///C:/b/abs_e2qn6c85jb/croot/notebook_1690985290943/work  
notebook_shim @ file:///C:/ci_311/notebook-shim_1678144850856/work        
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.6.1
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.0.0
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso==0.8.3
patsy==0.5.6
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
plotly==5.18.0
posthog==3.5.0
prometheus-client @ file:///C:/ci_311/prometheus_client_1679591942558/work
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work       
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent @ file:///C:/ci_311/pyrsistent_1676422695500/work
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
pywinpty @ file:///C:/ci_311/pywinpty_1677707791185/work/target/wheels/pywinpty-2.0.10-cp311-none-win_amd64.whl
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
rsa==4.9
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work     
simpervisor==1.0.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.1
sniffio @ file:///C:/ci_311/sniffio_1676425339093/work
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
spinners==0.0.24
SQLAlchemy==2.0.11
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.1
streamlit==1.30.0
sympy @ file:///C:/ci_311_rebuilds/sympy_1679009400182/work
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado @ file:///C:/ci_311/terminado_1678228513830/work
threadpoolctl==3.1.0
tinycss2 @ file:///C:/ci_311/tinycss2_1676425376744/work
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
ucimlrepo==0.0.3
urllib3==2.1.0
uvicorn==0.27.1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.8
webencodings==0.5.1
websocket-client @ file:///C:/ci_311/websocket-client_1676426063281/work  
websockets==12.0
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0